% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semana_santa.R
\name{semana_santa}
\alias{semana_santa}
\title{Determina si una fecha pertenece a la Semana Santa}
\usage{
semana_santa(fechas)
}
\arguments{
\item{fechas}{Un vector de fechas en formato `Date` que se desea verificar.}
}
\value{
Un vector lógico de la misma longitud que `fechas`, donde si la fecha corresponde a la Semana Santa es `TRUE` y `FALSE` si no.
}
\description{
Esta función recibe un vector de fechas y determina si cada fecha cae dentro de los rangos de Semana Santa (Domingo de Ramos a Domingo de Resurrección) indicando TRUE si la fecha pertenece a la Semana Santa y FALSE si no.
}
\examples{
# El usuario debe tener su propio data frame `df` con una columna de fechas, en formato ymd.
#df$ss <- semana_santa(df$fecha_evento)

}
