% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semana_santa_anio.R
\name{semana_santa_anio}
\alias{semana_santa_anio}
\title{Determina si una fecha pertenece a Semana Santa y su año}
\usage{
semana_santa_anio(fechas)
}
\arguments{
\item{fechas}{Un vector de fechas en formato `Date` que se desea verificar.}
}
\value{
Un vector de caracteres con dos tipos posibles de valores, ya sea FALSE, o bien, el texto "Semana Santa (AÑO)".
}
\description{
Esta función recibe un vector de fechas y determina si cada fecha cae dentro de los rangos de Semana Santa (Domingo de Ramos a Domingo de Resurrección) y el año al que pertenece de Semana Santa, ejemplo: Semana Santa (2019), si no pertenece a Semana Santa de ningun año, indica FALSE.
}
\examples{
# El usuario debe tener su propio data frame `df` con una columna de fechas, en formato ymd.
#df$ss_anio <- semana_santa_anio(df$fecha_evento)

}
