% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semana_santa_dia.R
\name{semana_santa_dia}
\alias{semana_santa_dia}
\title{Determina el día de la semana mayor, solo si la fecha pertenece a Semana Santa}
\usage{
semana_santa_dia(fechas)
}
\arguments{
\item{fechas}{Un vector de fechas en formato `Date` que se desea verificar.}
}
\value{
Un factor ordenado según la secuencia cronológica de la Semana Santa (de Domingo de Ramos a Domingo de Resurrección), no incluye los acentos, ya que el data frame original evita el uso de acentos al ser nombres de columnas, el usuario puede hacer los cambios de forma normal una vez teniendo el resultado.
}
\description{
Esta función recibe un vector de fechas, determina si cada fecha cae dentro de los rangos de Semana Santa (Domingo de Ramos a Domingo de Resurrección) y edevuelve el día de la semana mayor a la que pertenece, ejemplo: "Viernes Santo". Si la fecha no pertenece a Semana Santa, el resultado es NA.
}
\examples{
# El usuario debe tener su propio data frame `df` con una columna de fechas, en formato ymd.
#df$ss_dia <- semana_santa_dia(df$fecha_evento)

}
